requirements.txt
altgraph==0.17.4
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-humanize==0.1.2
djangorestframework==3.14.0
dnspython==2.4.2
eventlet==0.33.3
flake8==6.1.0
greenlet==3.0.2
humanize==4.9.0
idna==3.6
importlib-metadata==7.0.0
isort==5.12.0
kombu==5.3.4
mccabe==0.7.0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
zipp==3.17.0
